change - Apartment controller

+++ app/Http/Controllers/Admin/ApartmentController.php
@@ -12,6 +12,8 @@ use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
 use Illuminate\Support\Facades\Http;
+use illuminate\support\Facades\Gate;
+
 
 class ApartmentController extends Controller
 {
@@ -20,6 +22,7 @@ class ApartmentController extends Controller
      */
     public function index()
     {
+
         $current_user = Auth::id();
         if ($current_user == '1') {
             $apartments = Apartment::paginate(16);
@@ -108,12 +111,17 @@ class ApartmentController extends Controller
      */
     public function show(Apartment $apartment)
     {
-        $images = Image::where('apartment_id', $apartment->id)->get();
+        // Validazione id utente tramite policy
 
-        if (Auth::id() == $apartment->user_id || Auth::id() == 1) {
+        $response = Gate::inspect('view', $apartment);
+
+        if ($response->allowed()) {
+
+            $images = Image::where('apartment_id', $apartment->id)->get();
+
             return view('admin.apartments.show', compact('apartment', 'images'));
         } else {
-            return redirect()->route('admin.apartments.index');
+            abort(403);
         }
     }
 
@@ -122,10 +130,19 @@ class ApartmentController extends Controller
      */
     public function edit(Apartment $apartment)
     {
-        $services = Service::orderBy('name', 'ASC')->get();
-        $images = Image::where('apartment_id', $apartment->id)->get();
+        // Validazione id utente tramite policy
 
-        return view('admin.apartments.edit', compact('apartment', 'services', 'images'));
+        $response = Gate::inspect('update', $apartment);
+
+        if ($response->allowed()) {
+
+            $services = Service::orderBy('name', 'ASC')->get();
+            $images = Image::where('apartment_id', $apartment->id)->get();
+
+            return view('admin.apartments.edit', compact('apartment', 'services', 'images'));
+        } else {
+            abort(403);
+        }
     }
 
     /**
@@ -133,59 +150,68 @@ class ApartmentController extends Controller
      */
     public function update(UpdateApartmentRequest $request, Apartment $apartment)
     {
-        $request->validate([
-            'name' => 'required',
-            'description' => 'required|max:500',
-            'rooms' => 'required|numeric|gt:0',
-            'beds' => 'required|numeric|gt:0',
-            'bathrooms' => 'required|numeric|gt:0',
-            'square_meters' => 'required|numeric|gt:0',
-            'cover_image' => 'file|max:2048',
-            'services' => 'required|min:1'
-        ]);
+        // Validazione id utente tramite policy
 
-        $data = $request->all();
+        $response = Gate::inspect('update', $apartment);
 
-        if ($request->hasFile('cover_image')) {
-            $path = Storage::put('cover_images', $request->cover_image);
-            $data['cover_image'] = $path;
+        if ($response->allowed()) {
 
-            if ($apartment->cover_image) {
-                Storage::delete($apartment->cover_image);
+            $request->validate([
+                'name' => 'required',
+                'description' => 'required|max:500',
+                'rooms' => 'required|numeric|gt:0',
+                'beds' => 'required|numeric|gt:0',
+                'bathrooms' => 'required|numeric|gt:0',
+                'square_meters' => 'required|numeric|gt:0',
+                'cover_image' => 'file|max:2048',
+                'services' => 'required|min:1'
+            ]);
+
+            $data = $request->all();
+
+            if ($request->hasFile('cover_image')) {
+                $path = Storage::put('cover_images', $request->cover_image);
+                $data['cover_image'] = $path;
+
+                if ($apartment->cover_image) {
+                    Storage::delete($apartment->cover_image);
+                }
             }
-        }
 
-        $apartment->update($data);
+            $apartment->update($data);
 
-        if ($request->hasFile('images')) {
-
+            if ($request->hasFile('images')) {
 
 
-            $images = $request->images;
-            foreach ($images as $image) {
-                $link = Storage::put('images', $image);
-                $current_image['link'] = $link;
-                $current_image['apartment_id'] = $apartment->id;
-                $new_image = Image::create($current_image);
+
+                $images = $request->images;
+                foreach ($images as $image) {
+                    $link = Storage::put('images', $image);
+                    $current_image['link'] = $link;
+                    $current_image['apartment_id'] = $apartment->id;
+                    $new_image = Image::create($current_image);
+                }
             }
-        }
 
-        if ($request->has('old_images')) {
-            $old_images = $request->old_images;
-            foreach ($old_images as $old_image) {
-                $current_image = json_decode($old_image);
-                Storage::delete($current_image->link);
-                Image::destroy($current_image->id);
+            if ($request->has('old_images')) {
+                $old_images = $request->old_images;
+                foreach ($old_images as $old_image) {
+                    $current_image = json_decode($old_image);
+                    Storage::delete($current_image->link);
+                    Image::destroy($current_image->id);
+                }
             }
-        }
 
-        if ($request->has('services')) {
-            $apartment->services()->sync($data['services']);
+            if ($request->has('services')) {
+                $apartment->services()->sync($data['services']);
+            } else {
+                $apartment->services()->sync([]);
+            }
+
+            return redirect()->route('admin.apartments.show', $apartment);
         } else {
-            $apartment->services()->sync([]);
+            abort(403);
         }
-
-        return redirect()->route('admin.apartments.show', $apartment);
     }
 
     /**
